Change routes

+++ app/services/users.py
@@ -1,8 +1,10 @@
 import random
 import string
 
+from fastapi import HTTPException
 from passlib.context import CryptContext
 
+from app.conf.messages import ERROR_USER_NOT_FOUND
 from app.schemas.user_schemas import SignUpRequestModel, UserUpdate
 from app.utils.repository import AbstractRepository
 
@@ -22,14 +24,24 @@ class UsersService:
 
     async def get_user_by_id(self, user_id: int):
         filter_by = {"id": user_id}
+        user = await self.users_repo.find_by_filter(filter_by)
+        if user is None:
+            raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         return await self.users_repo.find_by_filter(filter_by)
 
     async def update_user(self, user_id: int, user_data: UserUpdate):
         user_dict = user_data.model_dump()
         filter_by = {"id": user_id}
+        user = await self.users_repo.find_by_filter(filter_by)
+        if user is None:
+            raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         return await self.users_repo.update_by_filter(filter_by, user_dict)
 
     async def delete_user(self, user_id: int):
+        filter_by = {"id": user_id}
+        user = await self.users_repo.find_by_filter(filter_by)
+        if user is None:
+            raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         await self.users_repo.delete_by_id(user_id)
 
     async def find_user_by_email(self, email: str):
